Flesh out footer with district contact information and logo
Build out footer region with logo.

Create custom block that loads the district for the page being viewed to
display the contact information for use in the footer. Currently coupled
with the module routing logic (pulls the route parameter to load the
district). This was the only way we could find to do it that seemed to
work.