Swap custom controller for autopathing module
The district content type routing was largely just to make the URLs
prettier which this module also accomplishes with the added advantage of
providing the node as the route parameter and alleviating the need to
perform multiple lookups to load the district to find the content node
as well as in the custom contact block.

Also start committing site configuration files (which include the
configuration of the autopath routes).